scanner/requirements.txt
appwrite
requests
python-dotenv
schedule

